repo: refactor the repositories
Split up the repo.py into a separate module, having multiple
submodule (base, local, remote). Right now the BaseRepo just
handles the _id assignment.
