+++ lib/bup/repo/base.py
@@ -1,0 +1,28 @@
+
+from __future__ import absolute_import
+
+_next_repo_id = 0
+_repo_ids = {}
+
+def _repo_id(key):
+    global _next_repo_id, _repo_ids
+    repo_id = _repo_ids.get(key)
+    if repo_id:
+        return repo_id
+    next_id = _next_repo_id = _next_repo_id + 1
+    _repo_ids[key] = next_id
+    return next_id
+
+class BaseRepo(object):
+    def __init__(self, key, compression_level=None,
+                 max_pack_size=None, max_pack_objects=None):
+        self._id = _repo_id(key)
+        self.compression_level = compression_level
+        self.max_pack_size = max_pack_size
+        self.max_pack_objects = max_pack_objects
+
+    def id(self):
+        """Return an identifier that differs from any other repository that
+        doesn't share the same repository-specific information
+        (e.g. refs, tags, etc.)."""
+        return self._id
