+++ lib/bup/repo/local.py
@@ -4,35 +4,24 @@ import os, subprocess
 from os.path import realpath
 from functools import partial
 
-from bup import client, git, vfs
+from bup import git, vfs
+from bup.repo.base import BaseRepo
 
 
-_next_repo_id = 0
-_repo_ids = {}
-
-def _repo_id(key):
-    global _next_repo_id, _repo_ids
-    repo_id = _repo_ids.get(key)
-    if repo_id:
-        return repo_id
-    next_id = _next_repo_id = _next_repo_id + 1
-    _repo_ids[key] = next_id
-    return next_id
-
-class LocalRepo:
+class LocalRepo(BaseRepo):
     def __init__(self, repo_dir=None, compression_level=None,
                  max_pack_size=None, max_pack_objects=None,
                  objcache_maker=None):
         self._packwriter = None
         self.repo_dir = realpath(git.guess_repo(repo_dir))
+        super(LocalRepo, self).__init__(self.repo_dir,
+                                        compression_level=compression_level,
+                                        max_pack_size=max_pack_size,
+                                        max_pack_objects=max_pack_objects)
         self._cp = git.cp(self.repo_dir)
         self.rev_list = partial(git.rev_list, repo_dir=self.repo_dir)
         self.config = partial(git.git_config_get, repo_dir=self.repo_dir)
-        self._id = _repo_id(self.repo_dir)
         self._dumb_server_mode = None
-        self.compression_level = compression_level
-        self.max_pack_size = max_pack_size
-        self.max_pack_objects = max_pack_objects
         self.objcache_maker = objcache_maker
 
     @classmethod
@@ -60,12 +49,6 @@ class LocalRepo:
             self._dumb_server_mode = os.path.exists(git.repo(b'bup-dumb-server',
                                                              repo_dir=self.repo_dir))
         return self._dumb_server_mode
-
-    def id(self):
-        """Return an identifier that differs from any other repository that
-        doesn't share the same repository-specific information
-        (e.g. refs, tags, etc.)."""
-        return self._id
 
     def is_remote(self):
         return False
@@ -178,131 +161,6 @@ class LocalRepo:
             w = self._packwriter
             self._packwriter = None
             return w.close(run_midx=run_midx)
-
-    def abort_writing(self):
-        if self._packwriter:
-            self._packwriter.abort()
-
-def make_repo(address, create=False, compression_level=None,
-              max_pack_size=None, max_pack_objects=None):
-    return RemoteRepo(address, create=create,
-                      compression_level=compression_level,
-                      max_pack_size=max_pack_size,
-                      max_pack_objects=max_pack_objects)
-
-class RemoteRepo:
-    def __init__(self, address, create=False, compression_level=None,
-                 max_pack_size=None, max_pack_objects=None):
-        # if client.Client() raises an exception, have a client
-        # anyway to avoid follow-up exceptions from __del__
-        self.client = None
-        self.client = client.Client(address)
-        self.rev_list = self.client.rev_list
-        self.config = self.client.config
-        self.list_indexes = self.client.list_indexes
-        self.read_ref = self.client.read_ref
-        self.send_index = self.client.send_index
-        self.join = self.client.join
-        self.refs = self.client.refs
-        self.resolve = self.client.resolve
-        self._id = _repo_id(address)
-        self._packwriter = None
-        self.compression_level = compression_level
-        self.max_pack_size = max_pack_size
-        self.max_pack_objects = max_pack_objects
-
-    def close(self):
-        self.finish_writing()
-        if self.client:
-            self.client.close()
-            self.client = None
-
-    def __del__(self):
-        self.close()
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, type, value, traceback):
-        self.close()
-
-    def id(self):
-        """Return an identifier that differs from any other repository that
-        doesn't share the same repository-specific information
-        (e.g. refs, tags, etc.)."""
-        return self._id
-
-    def update_ref(self, refname, newval, oldval):
-        self.finish_writing()
-        return self.client.update_ref(refname, newval, oldval)
-
-    def _ensure_packwriter(self):
-        if not self._packwriter:
-            self._packwriter = self.client.new_packwriter(
-                                    compression_level=self.compression_level,
-                                    max_pack_size=self.max_pack_size,
-                                    max_pack_objects=self.max_pack_objects)
-
-    def is_remote(self):
-        return True
-
-    def cat(self, ref):
-        """If ref does not exist, yield (None, None, None).  Otherwise yield
-        (oidx, type, size), and then all of the data associated with
-        ref.
-
-        """
-        # Yield all the data here so that we don't finish the
-        # cat_batch iterator (triggering its cleanup) until all of the
-        # data has been read.  Otherwise we'd be out of sync with the
-        # server.
-        items = self.client.cat_batch((ref,))
-        oidx, typ, size, it = info = next(items)
-        yield info[:-1]
-        if oidx:
-            for data in it:
-                yield data
-        assert not next(items, None)
-
-    def write_commit(self, tree, parent,
-                     author, adate_sec, adate_tz,
-                     committer, cdate_sec, cdate_tz,
-                     msg):
-        self._ensure_packwriter()
-        return self._packwriter.new_commit(tree, parent,
-                                           author, adate_sec, adate_tz,
-                                           committer, cdate_sec, cdate_tz,
-                                           msg)
-
-    def write_tree(self, shalist):
-        self._ensure_packwriter()
-        return self._packwriter.new_tree(shalist)
-
-    def write_data(self, data):
-        self._ensure_packwriter()
-        return self._packwriter.new_blob(data)
-
-    def write_symlink(self, target):
-        self._ensure_packwriter()
-        return self._packwriter.new_blob(target)
-
-    def write_bupm(self, data):
-        self._ensure_packwriter()
-        return self._packwriter.new_blob(data)
-
-    def just_write(self, sha, type, content):
-        self._ensure_packwriter()
-        return self._packwriter.just_write(sha, type, content)
-
-    def exists(self, sha, want_source=False):
-        self._ensure_packwriter()
-        return self._packwriter.exists(sha, want_source=want_source)
-
-    def finish_writing(self, run_midx=True):
-        if self._packwriter:
-            w = self._packwriter
-            self._packwriter = None
-            return w.close()
 
     def abort_writing(self):
         if self._packwriter:
